% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution.R
\name{coef}
\alias{coef}
\alias{coef.linreg}
\title{Coefficients}
\usage{
\method{coef}{linreg}(reg)
}
\arguments{
\item{reg}{object which return from linreg class}
}
\description{
Print the coefficients as a named vector.
}
