% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution.R
\name{plot}
\alias{plot}
\alias{plot.linreg}
\title{Plot data}
\usage{
\method{plot}{linreg}(reg)
}
\arguments{
\item{reg}{object which return from linreg class}
}
\description{
This function plot two graph using ggplot2
}
