% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution.R
\name{print}
\alias{print}
\alias{print.linreg}
\title{Print function}
\usage{
\method{print}{linreg}(reg)
}
\arguments{
\item{reg}{object which return from linreg class}
}
\value{
print the coefficient as lm function.
}
\description{
This function print the coefficient of regretion.
}
