% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\docType{class}
\name{ridgereg}
\alias{ridgereg}
\title{Ridge regrestion model}
\description{
RC type class for Ridge Regression model.
}
\section{Fields}{

\describe{
\item{\code{formula}}{linear formula as symbolic model to filter data.}

\item{\code{data}}{should be data frame}

\item{\code{lambda}}{should be a numeric}
}}


