% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution.R
\name{summery}
\alias{summery}
\alias{summary.linreg}
\title{Summery}
\usage{
\method{summary}{linreg}(reg)
}
\arguments{
\item{reg}{object which return from linreg class}
}
\description{
Print the coefficients with their standard error, t-value and p-value as well as the estimate of ˆσ and the degrees
 of freedom in the model, same as printed of lm function.
}
