% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumsq_cpp.R
\name{sumsq_cpp}
\alias{sumsq_cpp}
\title{Sum of squares via Rcpp}
\usage{
sumsq_cpp(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Numeric scalar
}
\description{
Sum of squares via Rcpp
}
\examples{
sumsq_cpp(1:5)

}
